Cast to u8

+++ src/utils.rs
@@ -4,5 +4,5 @@ pub fn get_terminator_position(buf : &[u8]) -> usize{
 
 
 pub fn to_bytes(v : u16) -> (u8, u8){
-    (v & 0x00FF, v >> 8)
+    ((v & 0x00FF) as u8, (v >> 8) as u8)
 }
